requirements.txt
python-telegram-bot[job-queue]
flask
requests
python-dotenv